+++ docs/release_notes.rst
@@ -10,6 +10,7 @@ Release notes for **Blogit**.
 
 * Add support for Django 2.0 & 2.1, drop support for Django 1.11 and below.
 * Drop support for Python 2.
+* Format html rendered in admin list.
 
 0.4.13
 ======